+++ src/test/kotlin/de/montanero/aoc2019/intcode/IntcodeFile.kt
@@ -2,13 +2,13 @@ package de.montanero.aoc2019.intcode
 
 object IntcodeFile {
 
-    fun read(resource: String): List<Int> {
+    fun read(resource: String): List<Long> {
         val reader = IntcodeFile::class.java.getResourceAsStream(resource).reader(Charsets.UTF_8)
-        val l: MutableList<Int> = mutableListOf()
+        val l: MutableList<Long> = mutableListOf()
         reader.use {
             it.forEachLine {
                 it.split(",").forEach {
-                    l.add(it.toInt())
+                    l.add(it.toLong())
                 }
             }
         }
